% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subdbr-package.R
\docType{package}
\name{subdbr-package}
\alias{subdbr}
\alias{subdbr-package}
\title{subdbr: ETL Package For Subscription Databases}
\description{
This package helps process data for subscription database metric.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Auckland-Council-CC-Insights-Team/subdbr}
  \item Report bugs at \url{https://github.com/Auckland-Council-CC-Insights-Team/subdbr/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
